Set initial deadline tolerance to 0
THIS IS A BREAKING CHANGE! Existing instances will be changed to have a
deadline tolerance of 0 hours. You need to set your desired deadline
tolerance again after applying this migration (if it is not 0).

This should have been part of 99e9da8908b3151b402012e43858efcb147c363c.
But I didn't know that this is required back then...