Make external directory configurable for each task
The string ${TASK_ID} will then be replaced with the specified custom id
of a task. If it is not given, the string for the external directory is
not altered.
Instead of having to tie the underlying ID of a task to the task title
(or some id generated by the database), we can specify a custom id that
is not visible to the user.
This id is also available as an environment variable within a checker as
TASK_ID_CUSTOM.