Use new STORAGES setting instead of DEFAULT_FILE_STORAGE
DEFAULT_FILE_STORAGE is deprecated since Django 4.2. As the new setting
is not merged with the default setting, we also need to specify the
storage engine for static files (previously defined with
STATICFILES_STORAGE).
Our own UploadStorage engine did nothing more than to override the
default values of location and base_url. For that reason, it is replaced
with the FileSystemStorage and the appropriate options.