Allow omitting the custom task ID
If an external directory contains a TASK_ID_CUSTOM variable that needs
to be replaced but no custom task ID is present, the external directory
will not be mounted. This allows having tasks that don't need an
external directory mounted (and have no fitting directory to mount)
while still being able to have other tasks that require an external
directory to be mounted.